Extract ReferenceExportService; split tests across router/service/task/integration
